+++ src/code_gen/test_haskell.rs
@@ -7,6 +7,7 @@ const FILENAME: &str = "code_gen_test.snow";
 
 fn code_gen_from_string(src: &str) -> CResult<String> {
     let tokens = scanner::scanner(FILENAME, src)?;
+    dbg!(&tokens);
     let (_, funcs) = match parser::parser().parse(&tokens) {
         Ok((t, f)) => (t, f),
         Err(t) => (t, FunctionList::new()),
@@ -26,7 +27,8 @@ fn hello_world() -> CResult<()> {
 #[test]
 fn hello_world_with_do_block() -> CResult<()> {
     let src = r#"main = do
-    println "Hello World""#;
+    println "Hello World"
+"#;
     let left = code_gen_from_string(src)?;
     let right = r#"main = do
     putStr "Hello World" 
@@ -39,7 +41,8 @@ fn hello_world_with_do_block() -> CResult<()> {
 fn hello_world_with_do_block_x_2() -> CResult<()> {
     let src = r#"main = do
     println "Hello World"
-    println "Hello World""#;
+    println "Hello World"
+"#;
     let left = code_gen_from_string(src)?;
     let right = r#"main = do
     putStr "Hello World"