+++ src/repl/mod.rs
@@ -1,4 +1,0 @@
-use super::excute_with_env_of;
-use super::function::FunctionList;
-mod shell;
-pub use shell::run;