+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Version 0.0.0-alpha-0.0.0
 
 - Created a custom lexer for parsing indenting and function definitions and calls.
+- Created a custom parser.
 - Started trans piling to Haskell rather than assembly to speed up development.
+- Created Error message display for user.